requirements.txt
Django>=4.2
djangorestframework
pandas
openpyxl
python-dotenv
django-cors-headers
gunicorn
